Cast it.

+++ src/linad99/sgradclc.cpp
@@ -186,8 +186,8 @@ void gradcalc(int nvar, const dvector& _g)
 #endif
 
    // back up the file one buffer size and read forward
-   off_t offset = sizeof(grad_stack_entry)
-                  * gradient_structure::GRAD_STACK1->length;
+   off_t offset = (off_t)(sizeof(grad_stack_entry)
+                  * gradient_structure::GRAD_STACK1->length);
    off_t lpos = lseek(gradient_structure::GRAD_STACK1->_GRADFILE_PTR,
       -offset,SEEK_CUR);
